Implement methods and constructor/initializer calls for classes.
The changes to concatenate() in 'vm.c' also fixed an earlier bug with string addition.
This was primarily since we passed too long of a length to makeString, which led to the null byte from the first string being included in the middle of our string sum(?), effectively truncating the string (since any printing operations stop at the null byte.